Rename namespases in the Lesson 6
